[#1023] Updated to use is_zaak_visible() instead of is_object_visible() to also check for internal/external cases
Also moved some functions to a better place
